NS1: Split provider to multiple files
Grew long, let's enhance readability a bit.
